==> analysis/3_Text_similarities_Figure_2B.Rmd <==
---
title: "Text_similarities"
author: "ZZ"
# author: "zuzannazagrodzka"
date: "2022-05-05"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

# Source of the functions: https://www.markhw.com/blog/word-similarity-graphs


We combined two methods to be able visually explore topics and words' connections in the aims and missions documents.
We used the word similarity method (https://www.markhw.com/blog/word-similarity-graphs) to calculate similarities between words and later to be able to plot them with topic modeling score values (package stm)

The three primary steps were followed: 
1. Calculating the similarities between words (Cosine matrix). 

2. Formatting these similarity scores into a symmetric matrix, where the diagonal contains 0s and the off-diagonal cells are similarities between words.

3. Clustering nodes  using a community detection algorithm (here: Walktrap algorithm) and plotting the data into circular hierarchical dendrograms for each of the stakeholder group separately (Figure 2B). 


# Libraries
``` {r}
# remotes::install_github('talgalili/dendextend')
library(dendextend)
library(igraph)
library(ggraph)
library(reshape2)
library(dplyr)
library(tidyverse)
library(tidytext)

```


# Below chunk contain functions which were used to create words similarities matrix. Source of the code: 
https://www.markhw.com/blog/word-similarity-graphs

``` {r}
# Cosine matrix
cosine_matrix <- function(tokenized_data, lower = 0, upper = 1, filt = 0) {
  
  if (!all(c("word", "id") %in% names(tokenized_data))) {
    stop("tokenized_data must contain variables named word and id")
  }
  
  if (lower < 0 | lower > 1 | upper < 0 | upper > 1 | filt < 0 | filt > 1) {
    stop("lower, upper, and filt must be 0 <= x <= 1")
  }
  docs <- length(unique(tokenized_data$id))
  
  out <- tokenized_data %>%
    count(id, word) %>%
    group_by(word) %>%
    mutate(n_docs = n()) %>%
    ungroup() %>%
    filter(n_docs <= (docs * upper) & n_docs > (docs * lower)) %>%
    select(-n_docs) %>%
    mutate(n = 1) %>%
    spread(word, n, fill = 0) %>%
    select(-id) %>%
    as.matrix() %>%
    lsa::cosine()
  
  filt <- quantile(out[lower.tri(out)], filt)
  out[out < filt] <- diag(out) <- 0
  out <- out[rowSums(out) != 0, colSums(out) != 0]
  
  return(out)
}


# Walktrap_topics

walktrap_topics <- function(g, ...) {
  wt <- igraph::cluster_walktrap(g, ...)
  
  membership <- igraph::cluster_walktrap(g, ...) %>% 
      igraph::membership() %>% 
      as.matrix() %>% 
      as.data.frame() %>% 
      rownames_to_column("word") %>% 
      arrange(V1) %>% 
      rename(group = V1)
  
  dendrogram <- stats::as.dendrogram(wt)
  
  return(list(membership = membership, dendrogram = dendrogram))
}
```



# Importing data
``` {r}
corpus_df <- read.csv("./output/created_datasets/dataset_words_stm_5topics.csv")
```

# Running above functions for each stakeholder seperately

```{r}
corpus_df <- read.csv("./output/created_datasets/dataset_words_stm_5topics.csv")
stake_names <- unique(corpus_df$stakeholder)
count <- 1
figure_list <- vector()

for (stake in stake_names) {
  n <- count # number of the stakeholder in the list
  stakeholder_name <- stake_names[n] # stakeholder's name
  
  # selecting data with the right name of the stakeholder
  dat <- corpus_df %>%
  select(stakeholder, sentence_doc, word) %>%
  rename(id = sentence_doc) %>% 
  filter(stakeholder %in% stakeholder_name) %>% 
  select(-stakeholder)
  ##################### 
  
  # Calculating similarity matrix
  cos_mat <- cosine_matrix(dat, lower = 0.050, upper = 1, filt = 0.9)
  
  # Getting words used in the network
  word_list_net <- rownames(cos_mat)
  grep("workflow", word_list_net)

  print(dim(cos_mat))
  
  # Creating a graph
  g <- graph_from_adjacency_matrix(cos_mat, mode = "undirected", weighted = TRUE)
  topics = walktrap_topics(g)
  print(stakeholder_name)
  plot(topics$dendrogram)

  subtrees <- partition_leaves(topics$dendrogram)
  leaves <- subtrees[[1]]  


  pathRoutes <- function(leaf) {
    which(sapply(subtrees, function(x) leaf %in% x))
  }

  paths <- lapply(leaves, pathRoutes)

  edges = NULL
  for(a in 1:length(paths)){
    # print(a)
    for(b in c(1:(length(paths[[a]])-1))){
     if(b == (length(paths[[a]])-1)){
        tmp_df = data.frame(
         from = paths[[a]][b],
         to = leaves[a]
       )
     } else {
        tmp_df = data.frame(
          from = paths[[a]][b],
          to = paths[[a]][b+1]
       )
     }
     edges = rbind(edges, tmp_df)
   }
  }


  connect = melt(cos_mat) # library reshape required
  colnames(connect) = c("from", "to", "value")
  connect = subset(connect, value != 0)

  # create a vertices data.frame. One line per object of our hierarchy
  vertices  <-  data.frame(
    name = unique(c(as.character(edges$from), as.character(edges$to))) 
  ) 
  # Let's add a column with the group of each name. It will be useful later to color points
  vertices$group  <-  edges$from[ match( vertices$name, edges$to ) ]


  #Let's add information concerning the label we are going to add: angle, horizontal adjustement and potential flip
  #calculate the ANGLE of the labels
  #vertices$id <- NA
  #myleaves <- which(is.na( match(vertices$name, edges$from) ))
  #nleaves <- length(myleaves)
  #vertices$id[ myleaves ] <- seq(1:nleaves)
  #vertices$angle <- 90 - 360 * vertices$id / nleaves

  # calculate the alignment of labels: right or left
  # If I am on the left part of the plot, my labels have currently an angle < -90
  #vertices$hjust <- ifelse( vertices$angle < -90, 1, 0)

  # flip angle BY to make them readable
  #vertices$angle <- ifelse(vertices$angle < -90, vertices$angle+180, vertices$angle)
  
  # replacing "group" value with the topic numer from the STM
  
  # Colour words by topic modelling topics (STM)
  # I have to change the group numbers in topics

  stm_top <- corpus_df %>%
    select(stakeholder, word, topic = topic) %>% 
    filter(stakeholder %in% stakeholder_name) %>%
    select(-stakeholder)
  

  stm_top$topic <- as.factor(stm_top$topic)
  stm_top <- unique(stm_top)


  # Removing group column and creating a new one based on the stm info
  vertices <- vertices

  vertices <- vertices %>% 
    left_join(stm_top, by = c("name" = "word")) %>% 
    select(-group) %>% 
    rename(group = topic)

  # Adding a beta value so I can use it in the graph
  stm_beta <- corpus_df %>% 
    select(stakeholder, word, beta = highest_mean_beta) %>% 
    filter(stakeholder %in% stakeholder_name) %>%
    select(-stakeholder)
  
  
  stm_beta <- unique(stm_beta)
  dim(stm_beta)

  vertices <- vertices %>% 
   left_join(stm_beta, by = c("name" = "word")) 


  vertices$beta_size = vertices$beta
  
  ##############################################################

  # Create a graph object
  mygraph <- igraph::graph_from_data_frame(edges, vertices=vertices)

  # The connection object must refer to the ids of the leaves:
  from  <-  match( connect$from, vertices$name)
  to  <-  match( connect$to, vertices$name)

  # Basic usual argument
  tmp_plt = ggraph(mygraph, layout = 'dendrogram', circular = T) + 
    geom_node_point(aes(filter = leaf, x = x*1.05, y=y*1.05, alpha=0.2), size =0, colour = "white") +
    #colour=group, size=value
          scale_colour_manual(values= c("dark green", "dark red", "darkgoldenrod", "dark blue", "black"), na.value = "black") +   
    geom_conn_bundle(data = get_con(from = from, to = to, col = connect$value), tension = 0.9, aes(colour = col, alpha = col+0.5), width = 1.1) +
    scale_edge_color_continuous(low="white", high="black") +
    # geom_node_text(aes(x = x*1.1, y=y*1.1, filter = leaf, label=name, angle = angle, hjust=hjust), size=5, alpha=1) +
    geom_node_text(aes(x = x*1.1, y=y*1.1, filter = leaf, label=name, colour = group, angle = 0, hjust = 0.5, size = beta_size),  alpha=1) +
    
     theme_void() +
    theme(
      legend.position="none",
      plot.margin=unit(c(0,0,0,0),"cm"),
    ) +
    expand_limits(x = c(-1.5, 1.5), y = c(-1.5, 1.5)) 

  g <- ggplot_build(tmp_plt)
  tmp_dat = g[[3]]$data
  tmp_dat$position = NA
  tmp_dat_a = subset(tmp_dat, x >= 0)
  tmp_dat_a = tmp_dat_a[order(-tmp_dat_a$y),]
  tmp_dat_a = subset(tmp_dat_a, !is.na(beta))
  tmp_dat_a$order = 1:nrow(tmp_dat_a)
  
  tmp_dat_b = subset(tmp_dat, x < 0)
  tmp_dat_b = tmp_dat_b[order(-tmp_dat_b$y),]
  tmp_dat_b = subset(tmp_dat_b, !is.na(beta))
  tmp_dat_b$order = (nrow(tmp_dat_a) +nrow(tmp_dat_b)) : (nrow(tmp_dat_a) +1)
  
  tmp_dat = rbind(tmp_dat_a, tmp_dat_b)
  
  vertices = left_join(vertices, tmp_dat[,c("name", "order")], by = c("name"))
  vert_save = vertices
  vertices = vert_save
  
  vertices$id <- NA
  vertices = vertices[order(vertices$order),]
  myleaves <- which(is.na( match(vertices$name, edges$from) ))
  nleaves <- length(myleaves)
  vertices$id[ myleaves ] <- seq(1:nleaves)
  vertices$angle <- 90 - 360 * vertices$id / nleaves
  
  # calculate the alignment of labels: right or left
  # If I am on the left part of the plot, my labels have currently an angle < -90
  vertices$hjust <- ifelse( vertices$angle < -90, 1, 0)
  
  # flip angle BY to make them readable
  vertices$angle <- ifelse(vertices$angle < -90, vertices$angle+180, vertices$angle)
  vertices = vertices[order(as.numeric(rownames(vertices))),]
  
  mygraph <- igraph::graph_from_data_frame(edges, vertices=vertices)
  
  figure_to_save <- ggraph(mygraph, layout = 'dendrogram', circular = T) + 
    geom_node_point(aes(filter = leaf, x = x*1.05, y=y*1.05, alpha=0.2, colour = group, size = beta_size)) +
    #colour=group, size=value
          scale_colour_manual(values= c("dark green", "dark red", "darkgoldenrod", "dark blue", "black"), na.value = "black") +   
    geom_conn_bundle(data = get_con(from = from, to = to, col = connect$value), tension = 0.99, aes(colour = col, alpha = col+0.5), width = 1.1) +
    scale_edge_color_continuous(low="grey90", high="black") +
    geom_node_text(aes(x = x*1.1, y=y*1.1, filter = leaf, label=name, colour = group, angle = angle, hjust = hjust, size = beta_size)) +
    
    theme_void() +
    theme(
      legend.position="none",
      plot.margin=unit(c(0,0,0,0),"cm"),
    ) +
    expand_limits(x = c(-1.5, 1.5), y = c(-1.5, 1.5))

  assign(paste0(stake_names[n], "_figure2B"), figure_to_save)
  # name_temp <- assign(paste0(stake_names[n], "_figure2b"), figure_to_save)
  figure_list <- append(figure_list, paste0(stake_names[n], "_figure2B"))
  
  print(figure_to_save)

  #####################
  count <-  count + 1 

}

```

```{r}
# Saving figures individually 

figure_list

figure_name <- paste0("./output/Figure_2B/advocates_5topics_figure2B.png")
png(file=figure_name,
  width= 3000, height= 3000, res=400)
  advocates_figure2B
  dev.off()


figure_name <- paste0("./output/Figure_2B/funders_5topics_figure2B.png")
png(file=figure_name,
  width= 3000, height= 3000, res=400)
  funders_figure2B
  dev.off()
  
figure_name <- paste0("./output/Figure_2B/journals_5topics_figure2B.png")
png(file=figure_name,
  width= 3000, height= 3000, res=400)
  journals_figure2B
  dev.off()
  
figure_name <- paste0("./output/Figure_2B/publishers_5topics_figure2B.png")
png(file=figure_name,
  width= 3000, height= 3000, res=400)
  publishers_figure2B
  dev.off()
  
figure_name <- paste0("./output/Figure_2B/repositories_5topics_figure2B.png")
png(file=figure_name,
  width= 3000, height= 3000, res=400)
  repositories_figure2B
  dev.off()
  
figure_name <- paste0("./output/Figure_2B/societies_5topics_figure2B.png")
png(file=figure_name,
  width= 3000, height= 3000, res=400)
  societies_figure2B
  dev.off()

```



# Additional analysis
Circular graph created on all documents (all stakeholders)

``` {r}
# selecting data with the right name of the stakeholder
dat <- corpus_df %>%
  select(sentence_doc, word) %>%
  rename(id = sentence_doc) 

  # Calculating similarity matrix
  cos_mat <- cosine_matrix(dat, lower = 0.035, upper = 1, filt = 0.9)
  
# Getting words used in the network
word_list_net <- rownames(cos_mat)
grep("workflow", word_list_net)

# Creating a graph
g <- graph_from_adjacency_matrix(cos_mat, mode = "undirected", weighted = TRUE)
topics = walktrap_topics(g)
plot(topics$dendrogram)

subtrees <- partition_leaves(topics$dendrogram)
leaves <- subtrees[[1]]  

pathRoutes <- function(leaf) {
  which(sapply(subtrees, function(x) leaf %in% x))
}

paths <- lapply(leaves, pathRoutes)

edges = NULL
for(a in 1:length(paths)){
  # print(a)
  for(b in c(1:(length(paths[[a]])-1))){
    if(b == (length(paths[[a]])-1)){
      tmp_df = data.frame(
        from = paths[[a]][b],
        to = leaves[a]
      )
    } else {
      tmp_df = data.frame(
        from = paths[[a]][b],
        to = paths[[a]][b+1]
      )
    }
    edges = rbind(edges, tmp_df)
  }
}
connect = melt(cos_mat)
colnames(connect) = c("from", "to", "value")
connect = subset(connect, value != 0)


# create a vertices data.frame. One line per object of our hierarchy
vertices  <-  data.frame(
  name = unique(c(as.character(edges$from), as.character(edges$to))) 
) 
# Let's add a column with the group of each name. It will be useful later to color points
vertices$group  <-  edges$from[ match( vertices$name, edges$to ) ]


#Let's add information concerning the label we are going to add: angle, horizontal adjustement and potential flip
#calculate the ANGLE of the labels
#vertices$id <- NA
#myleaves <- which(is.na( match(vertices$name, edges$from) ))
#nleaves <- length(myleaves)
#vertices$id[ myleaves ] <- seq(1:nleaves)
#vertices$angle <- 90 - 360 * vertices$id / nleaves

# calculate the alignment of labels: right or left
# If I am on the left part of the plot, my labels have currently an angle < -90
#vertices$hjust <- ifelse( vertices$angle < -90, 1, 0)

# flip angle BY to make them readable
#vertices$angle <- ifelse(vertices$angle < -90, vertices$angle+180, vertices$angle)

# replacing "group" value with the topic numer from the STM

# Colour words by topic modelling topics (STM)
# I have to change the group numbers in topics

stm_top <- corpus_df %>%
  select(word, topic = topic)

stm_top$topic <- as.factor(stm_top$topic)
stm_top <- unique(stm_top)

# Removing group column and creating a new one based on the stm info
vertices <- vertices

vertices <- vertices %>% 
  left_join(stm_top, by = c("name" = "word")) %>% 
  select(-group) %>% 
  rename(group = topic)

# Adding a beta value so I can use it in the graph
stm_beta <- corpus_df %>% 
  select(word, beta = highest_mean_beta)
stm_beta <- unique(stm_beta)
dim(stm_beta)

vertices <- vertices %>% 
  left_join(stm_beta, by = c("name" = "word")) 

vertices$beta_size = vertices$beta
##########################################################

# Create a graph object
mygraph <- igraph::graph_from_data_frame(edges, vertices=vertices)

# The connection object must refer to the ids of the leaves:
from  <-  match( connect$from, vertices$name)
to  <-  match( connect$to, vertices$name)

# Basic usual argument
tmp_plt = ggraph(mygraph, layout = 'dendrogram', circular = T) + 
  geom_node_point(aes(filter = leaf, x = x*1.05, y=y*1.05, alpha=0.2), size =0, colour = "white") +
  #colour=group, size=value
        scale_colour_manual(values= c("dark green", "dark red", "darkgoldenrod", "dark blue", "black"), na.value = "black") +   
  geom_conn_bundle(data = get_con(from = from, to = to, col = connect$value), tension = 0.9, aes(colour = col, alpha = col+0.5), width = 1.1) +
  scale_edge_color_continuous(low="white", high="black") +
  # geom_node_text(aes(x = x*1.1, y=y*1.1, filter = leaf, label=name, angle = angle, hjust=hjust), size=5, alpha=1) +
  geom_node_text(aes(x = x*1.1, y=y*1.1, filter = leaf, label=name, colour = group, angle = 0, hjust = 0.5, size = beta_size),  alpha=1) +
  
   theme_void() +
  theme(
    legend.position="none",
    plot.margin=unit(c(0,0,0,0),"cm"),
  ) +
  expand_limits(x = c(-1.5, 1.5), y = c(-1.5, 1.5)) 

g <- ggplot_build(tmp_plt)
tmp_dat = g[[3]]$data
tmp_dat$position = NA
tmp_dat_a = subset(tmp_dat, x >= 0)
tmp_dat_a = tmp_dat_a[order(-tmp_dat_a$y),]
tmp_dat_a = subset(tmp_dat_a, !is.na(beta))
tmp_dat_a$order = 1:nrow(tmp_dat_a)

tmp_dat_b = subset(tmp_dat, x < 0)
tmp_dat_b = tmp_dat_b[order(-tmp_dat_b$y),]
tmp_dat_b = subset(tmp_dat_b, !is.na(beta))
tmp_dat_b$order = (nrow(tmp_dat_a) +nrow(tmp_dat_b)) : (nrow(tmp_dat_a) +1)

tmp_dat = rbind(tmp_dat_a, tmp_dat_b)

vertices = left_join(vertices, tmp_dat[,c("name", "order")], by = c("name"))
vert_save = vertices
vertices = vert_save

vertices$id <- NA
vertices = vertices[order(vertices$order),]
myleaves <- which(is.na( match(vertices$name, edges$from) ))
nleaves <- length(myleaves)
vertices$id[ myleaves ] <- seq(1:nleaves)
vertices$angle <- 90 - 360 * vertices$id / nleaves

# calculate the alignment of labels: right or left
# If I am on the left part of the plot, my labels have currently an angle < -90
vertices$hjust <- ifelse( vertices$angle < -90, 1, 0)

# flip angle BY to make them readable
vertices$angle <- ifelse(vertices$angle < -90, vertices$angle+180, vertices$angle)
vertices = vertices[order(as.numeric(rownames(vertices))),]

mygraph <- igraph::graph_from_data_frame(edges, vertices=vertices)

figure_tosave <- ggraph(mygraph, layout = 'dendrogram', circular = T) + 
  geom_node_point(aes(filter = leaf, x = x*1.05, y=y*1.05, alpha=0.2, colour = group, size = beta_size)) +
  #colour=group, size=value
        scale_colour_manual(values= c("dark green", "dark red", "darkgoldenrod", "dark blue", "black"), na.value = "black") +   
  geom_conn_bundle(data = get_con(from = from, to = to, col = connect$value), tension = 0.99, aes(colour = col, alpha = col+0.5), width = 1.1) +
  scale_edge_color_continuous(low="grey90", high="black") +
  geom_node_text(aes(x = x*1.1, y=y*1.1, filter = leaf, label=name, colour = group, angle = angle, hjust = hjust, size = beta_size)) +
  
  theme_void() +
  theme(
    legend.position="none",
    plot.margin=unit(c(0,0,0,0),"cm"),
  ) +
  expand_limits(x = c(-1.6, 1.6), y = c(-1.6, 1.6))

figure_name <- paste0("./output/Other_figures/All_stakeholders_5topics_figure.png")
png(file=figure_name,
  width=3000, height=3000, res = 500)
  figure_tosave
  dev.off()
  
figure_tosave

```


# Session information
``` {r}
sessionInfo()
```
